+++ src/main/java/io/geekidea/springbootplus/SpringBootPlusApplication.java
@@ -22,6 +22,7 @@ import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.web.servlet.ServletComponentScan;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -39,6 +40,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 @EnableConfigurationProperties
 @EnableAdminServer
 @MapperScan({"io.geekidea.springbootplus.**.mapper"})
+@ServletComponentScan
 @SpringBootApplication
 public class SpringBootPlusApplication {