+++ src/main/java/io/geekidea/springbootplus/system/service/impl/SysUserServiceImpl.java
@@ -1,28 +1,12 @@
-/*
- * Copyright 2019-2029 geekidea(https://github.com/geekidea)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package io.geekidea.springbootplus.system.service.impl;
 
 import io.geekidea.springbootplus.system.entity.SysUser;
 import io.geekidea.springbootplus.system.mapper.SysUserMapper;
 import io.geekidea.springbootplus.system.service.SysUserService;
-import io.geekidea.springbootplus.system.web.param.SysUserQueryParam;
-import io.geekidea.springbootplus.system.web.vo.SysUserQueryVo;
+import io.geekidea.springbootplus.system.param.SysUserQueryParam;
+import io.geekidea.springbootplus.system.vo.SysUserQueryVo;
 import io.geekidea.springbootplus.common.service.impl.BaseServiceImpl;
-import io.geekidea.springbootplus.common.web.vo.Paging;
+import io.geekidea.springbootplus.common.vo.Paging;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -40,7 +24,7 @@ import java.io.Serializable;
  * </p>
  *
  * @author geekidea
- * @since 2019-10-05
+ * @since 2019-10-11
  */
 @Slf4j
 @Service
@@ -48,17 +32,17 @@ import java.io.Serializable;
 public class SysUserServiceImpl extends BaseServiceImpl<SysUserMapper, SysUser> implements SysUserService {
 
     @Autowired
-    private SysUserMapper SysUserMapper;
+    private SysUserMapper sysUserMapper;
 
     @Override
     public SysUserQueryVo getSysUserById(Serializable id) throws Exception {
-        return SysUserMapper.getSysUserById(id);
+        return sysUserMapper.getSysUserById(id);
     }
 
     @Override
-    public Paging<SysUserQueryVo> getSysUserPageList(SysUserQueryParam SysUserQueryParam) throws Exception {
-        Page page = setPageParam(SysUserQueryParam, OrderItem.desc("create_time"));
-        IPage<SysUserQueryVo> iPage = SysUserMapper.getSysUserPageList(page, SysUserQueryParam);
+    public Paging<SysUserQueryVo> getSysUserPageList(SysUserQueryParam sysUserQueryParam) throws Exception {
+        Page page = setPageParam(sysUserQueryParam, OrderItem.desc("create_time"));
+        IPage<SysUserQueryVo> iPage = sysUserMapper.getSysUserPageList(page, sysUserQueryParam);
         return new Paging(iPage);
     }