+++ src/main/java/io/geekidea/springbootplus/system/vo/SysUserQueryVo.java
@@ -1,0 +1,52 @@
+package io.geekidea.springbootplus.system.vo;
+
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import lombok.Data;
+import lombok.experimental.Accessors;
+import java.io.Serializable;
+
+import java.util.Date;
+
+/**
+ * <p>
+ * SystemUser 查询结果对象
+ * </p>
+ *
+ * @author geekidea
+ * @date 2019-10-11
+ */
+@Data
+@Accessors(chain = true)
+@ApiModel(value = "SysUserQueryVo对象", description = "SystemUser查询参数")
+public class SysUserQueryVo implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+    @ApiModelProperty(value = "主键")
+    private Long id;
+
+    @ApiModelProperty(value = "用户名")
+    private String username;
+
+    @ApiModelProperty(value = "昵称")
+    private String nickname;
+
+    @ApiModelProperty(value = "密码")
+    private String password;
+
+    @ApiModelProperty(value = "盐值")
+    private String salt;
+
+    @ApiModelProperty(value = "remark")
+    private String remark;
+
+    @ApiModelProperty(value = "状态，0：禁用，1：启用")
+    private Integer status;
+
+    @ApiModelProperty(value = "创建时间")
+    private Date createTime;
+
+    @ApiModelProperty(value = "修改时间")
+    private Date updateTime;
+
+}